+++ src/components/CarouselSlide.vue
